requirements.txt
invoke==1.2.0
attrdict==2.0.0
certifi==2018.11.29
chardet==3.0.4
cytoolz==0.9.0.1
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-typing==2.0.0
eth-utils==1.4.1
hexbytes==0.1.0
idna==2.8
lru-dict==1.1.6
parsimonious==0.8.1
py-solc==3.2.0
py-solc-simple==0.0.13
pycryptodome==3.7.2
requests==2.21.0
rlp==1.0.3
semantic-version==2.6.0
six==1.12.0
toolz==0.9.0
urllib3==1.24.1
web3==4.8.2
websockets==6.0
